Consolidate use of get_repo_for_repository() and clone_repository()

+++ lib/galaxy/webapps/tool_shed/controllers/upload.py
@@ -5,7 +5,6 @@ import tarfile
 import tempfile
 
 import requests
-from mercurial import commands
 
 from galaxy import (
     util,
@@ -43,7 +42,7 @@ class UploadController(BaseUIController):
         repository_id = kwd.get('repository_id', '')
         repository = repository_util.get_repository_in_tool_shed(trans.app, repository_id)
         repo_dir = repository.repo_path(trans.app)
-        repo = hg_util.get_repo_for_repository(trans.app, repository=None, repo_path=repo_dir, create=False)
+        repo = hg_util.get_repo_for_repository(trans.app, repo_path=repo_dir)
         uncompress_file = util.string_as_bool(kwd.get('uncompress_file', 'true'))
         remove_repo_files_not_in_tar = util.string_as_bool(kwd.get('remove_repo_files_not_in_tar', 'true'))
         uploaded_file = None
@@ -65,11 +64,9 @@ class UploadController(BaseUIController):
                 # Use mercurial clone to fetch repository, contents will then be copied over.
                 uploaded_directory = tempfile.mkdtemp()
                 repo_url = 'http%s' % url[len('hg'):]
-                repo_url = repo_url.encode('ascii', 'replace')
-                try:
-                    commands.clone(hg_util.get_configured_ui(), repo_url, uploaded_directory)
-                except Exception as e:
-                    message = 'Error uploading via mercurial clone: %s' % basic_util.to_html_string(str(e))
+                cloned_ok, error_message = hg_util.clone_repository(repo_url, uploaded_directory)
+                if not cloned_ok:
+                    message = 'Error uploading via mercurial clone: %s' % error_message
                     status = 'error'
                     basic_util.remove_dir(uploaded_directory)
                     uploaded_directory = None
